+++ src/RcsbFvStructure/StructureViewerInterface.ts
@@ -54,7 +54,6 @@ export interface ViewerCallbackManagerInterface {
 }
 
 export interface ViewerActionManagerInterface<R,L> {
-    //load<Z extends R|R[]>(loadConfig: Z): Z extends R ? Promise<L|undefined> : Promise<(L|undefined)[]>;
     load(loadConfig: R): Promise<L|undefined>;
     load(loadConfig: R[]): Promise<(L|undefined)[]>;
     removeStructure(removeConfig: R|Array<R>): Promise<void>;